backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.1.7
django-browser-reload==1.18.0
django-cors-headers==4.7.0
django-filter==25.1
django-jalali==7.3.0
django-jazzmin==3.0.1
django-tailwind==3.8.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.9
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.2.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
loguru==0.7.3
Markdown==3.7
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
python-dotenv==1.1.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
